pyproject.toml
[project]
name = "chorus-deck"
version = "0.1.0"
description = "Tool to generate slide decks with church songs and responses"
authors = [
    {name = "Margarida Silva", email = "anamargarida.eng@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-pptx (>=1.0.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "chorus_deck", from = "src"}]

[tool.poetry.scripts]
chorus-deck = "chorus_deck.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
